+++ program/system/system.h
@@ -6,7 +6,7 @@
 #define MINIFOC_SYSTEM_SYSTEM_H_
 
 void systick_config(void);
-void delayms(uint32_t count);
+void delayms(unsigned long count);
 void delay_decrement(void);
 void user_memset(void *buf, unsigned char data, unsigned char num);
 unsigned short float_to_int16(float data0);
